feat: add new `.tailDisplay()` function
