+++ deno.json
@@ -32,8 +32,8 @@
   ],
   "exports": "./mod.ts",
   "imports": {
-    "@david/console-static-text": "jsr:@david/console-static-text@^0.3.1",
-    "@david/shell": "jsr:@david/shell@^0.2",
+    "@david/console-static-text": "jsr:@david/console-static-text@^0.3.2",
+    "@david/shell": "jsr:@david/shell@^0.3",
     "@deno/dnt": "jsr:@deno/dnt@^0.42.3",
     "@david/path": "jsr:@david/path@^0.3.1",
     "@std/assert": "jsr:@std/assert@1",
